add conditional for local signing code

+++ build-scripts/macos.sh
@@ -12,23 +12,25 @@ if [ -f ".env" ]; then
   . .env
 fi
 
-security create-keychain -p "$KEYCHAIN_PASSWORD" "$KEYCHAIN_NAME" || echo "already exists"
+if [ -n "$LOCAL_BUILD" ]; then
+  security create-keychain -p "$KEYCHAIN_PASSWORD" "$KEYCHAIN_NAME" || echo "already exists"
 
-KEYCHAIN_NAME="login.keychain"
+  KEYCHAIN_NAME="login.keychain"
 
-certificate_path="/tmp/certificate.p12"
-echo "$SIGNING_CERTIFICATE_P12_DATA" | base64 -d > $certificate_path
-security default-keychain -d user -s "$KEYCHAIN_NAME"
+  certificate_path="/tmp/certificate.p12"
+  echo "$SIGNING_CERTIFICATE_P12_DATA" | base64 -d > $certificate_path
+  security default-keychain -d user -s "$KEYCHAIN_NAME"
 
-# security unlock-keychain -p "$KEYCHAIN_PASSWORD" "$KEYCHAIN_NAME"
-security import "$certificate_path" -f pkcs12 -k "$KEYCHAIN_NAME" -P "$SIGNING_CERTIFICATE_PASSWORD" -T /usr/bin/codesign -x 
-rm "$certificate_path"
-# security set-key-partition-list -S apple-tool:,apple: -s -k "$KEYCHAIN_PASSWORD" "$KEYCHAIN_NAME"
+  # security unlock-keychain -p "$KEYCHAIN_PASSWORD" "$KEYCHAIN_NAME"
+  security import "$certificate_path" -f pkcs12 -k "$KEYCHAIN_NAME" -P "$SIGNING_CERTIFICATE_PASSWORD" -T /usr/bin/codesign -x 
+  rm "$certificate_path"
+  # security set-key-partition-list -S apple-tool:,apple: -s -k "$KEYCHAIN_PASSWORD" "$KEYCHAIN_NAME"
 
-identity=$(security find-identity -v -p codesigning | grep -o "Developer ID Application.*(${NOTARIZE_TEAM_ID})")
-export CODESIGNING_IDENTITY="$identity"
+  identity=$(security find-identity -v -p codesigning | grep -o "Developer ID Application.*(${NOTARIZE_TEAM_ID})")
+  export CODESIGNING_IDENTITY="$identity"
 
-security set-keychain-settings -lut 1200
+  security set-keychain-settings -lut 1200
+fi
 
 export BUILD_DIR=build
 export IS_HEADLESS=0
@@ -115,51 +117,63 @@ cp -r "${BUILD_DIR}/themes/themes" "${BUNDLE_DIR}/CodeWhisperer.app/Contents/Res
 BUNDLE_PATH="${BUNDLE_DIR}/CodeWhisperer.app"
 cp -r "$BUNDLE_PATH" "$BUILD_DIR"
 
-codesign -v --timestamp --force --strict --options=runtime -s "$CODESIGNING_IDENTITY" -i io.fig.cli "$BUNDLE_PATH/Contents/MacOS/cw"
-codesign -v --timestamp --force --strict --options=runtime -s "$CODESIGNING_IDENTITY" -i io.fig.figterm "$BUNDLE_PATH/Contents/MacOS/cwterm" 
-codesign -v --timestamp --force --strict --options=runtime -s "$CODESIGNING_IDENTITY" -i io.fig.figterm "$BUNDLE_PATH/Contents/Helpers/FigInputMethod.app" 
-codesign -v --timestamp --force --strict --options=runtime -s "$CODESIGNING_IDENTITY" "$BUNDLE_PATH"
-codesign --verify --verbose --strict --entitlements entitlements.plist "$BUNDLE_PATH"
+if [ -n "$LOCAL_BUILD" ]; then
+  codesign -v --timestamp --force --strict --options=runtime -s "$CODESIGNING_IDENTITY" -i io.fig.cli "$BUNDLE_PATH/Contents/MacOS/cw"
+  codesign -v --timestamp --force --strict --options=runtime -s "$CODESIGNING_IDENTITY" -i io.fig.figterm "$BUNDLE_PATH/Contents/MacOS/cwterm" 
+  codesign -v --timestamp --force --strict --options=runtime -s "$CODESIGNING_IDENTITY" -i io.fig.figterm "$BUNDLE_PATH/Contents/Helpers/FigInputMethod.app" 
+  codesign -v --timestamp --force --strict --options=runtime -s "$CODESIGNING_IDENTITY" "$BUNDLE_PATH"
+  codesign --verify --verbose --strict --entitlements entitlements.plist "$BUNDLE_PATH"
 
-ditto -c -k --keepParent "$BUNDLE_PATH" Cw.zip
-xcrun notarytool submit Cw.zip --apple-id "$NOTARIZE_USERNAME" --password "$NOTARIZE_PASSWORD" --team-id "$NOTARIZE_TEAM_ID" --wait
-rm -f Cw.zip
-xcrun stapler staple "$BUNDLE_PATH"
-# Verify notarization ticket
-spctl -a -v "$BUNDLE_PATH"
+  ditto -c -k --keepParent "$BUNDLE_PATH" Cw.zip
+  xcrun notarytool submit Cw.zip --apple-id "$NOTARIZE_USERNAME" --password "$NOTARIZE_PASSWORD" --team-id "$NOTARIZE_TEAM_ID" --wait
+  rm -f Cw.zip
+  xcrun stapler staple "$BUNDLE_PATH"
+  # Verify notarization ticket
+  spctl -a -v "$BUNDLE_PATH"
+fi
 
 
-FILE_CONTENTS=$(jq -n \
-  --arg identity "$CODESIGNING_IDENTITY" \
-  --arg bundle "$BUNDLE_PATH" \
-  '{
-    "title": "CodeWhisperer",
-    "icon": "VolumeIcon.icns",
-    "background": "background.png",
-    "icon-size": 160,
-    "format": "ULFO",
-    "window": {
-      "size": {
-        "width": 660,
-        "height": 400
-      }
+FILE_JSON='{
+  "title": "CodeWhisperer",
+  "icon": "VolumeIcon.icns",
+  "background": "background.png",
+  "icon-size": 160,
+  "format": "ULFO",
+  "window": {
+    "size": {
+      "width": 660,
+      "height": 400
+    }
+  },
+  "contents": [
+    {
+      "x": 180,
+      "y": 170,
+      "type": "file",
+      "path": $bundle
     },
-    "contents": [
-      {
-        "x": 180,
-        "y": 170,
-        "type": "file",
-        "path": $bundle
-      },
-      {
-        "x": 480,
-        "y": 170,
-        "type": "link",
-        "path": "/Applications"
-      }
-    ]
-  }'
-)
+    {
+      "x": 480,
+      "y": 170,
+      "type": "link",
+      "path": "/Applications"
+    }
+  ]
+}'
+
+if [ -n "$LOCAL_BUILD" ]; then
+  FILE_CONTENTS=$(jq -n \
+    --arg identity "$CODESIGNING_IDENTITY" \
+    --arg bundle "$BUNDLE_PATH" \
+    "$FILE_JSON"
+  )
+else 
+  FILE_CONTENTS=$(jq -n \
+    --arg bundle "$BUNDLE_PATH" \
+    "$FILE_JSON"
+  )
+fi
+
 
 SPEC_FILE="bundle/dmg/spec.json"
 DMG="$BUILD_DIR/CodeWhisperer.dmg"
@@ -169,5 +183,7 @@ rm -f "$DMG"
 pnpm appdmg "$SPEC_FILE" "$DMG"
 rm "$SPEC_FILE"
 
-xcrun notarytool submit "$DMG" --apple-id "$NOTARIZE_USERNAME" --password "$NOTARIZE_PASSWORD" --team-id "$NOTARIZE_TEAM_ID" --wait
-spctl -a -t open --context context:primary-signature -v "$DMG"
+if [ -n "$LOCAL_BUILD" ]; then
+  xcrun notarytool submit "$DMG" --apple-id "$NOTARIZE_USERNAME" --password "$NOTARIZE_PASSWORD" --team-id "$NOTARIZE_TEAM_ID" --wait
+  spctl -a -t open --context context:primary-signature -v "$DMG"
+fi
